move logic from config to lib
