+++ app/controllers/v0/admin_controller.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'admin/postgres_check'
+
 module V0
   class AdminController < ApplicationController
     service_tag 'platform-base'
@@ -9,7 +11,7 @@ module V0
       app_status = {
         git_revision: AppInfo::GIT_REVISION,
         db_url: nil,
-        postgres_up: AppInfo.postgres_up
+        postgres_up: DatabaseHealthChecker.postgres_up
       }
       render json: app_status
     end
